fix: QR코드와 입금 링크 쿼리 오류 수정
- 기존 하나의 컴포넌트를 2개로 나누었습니다.
- 티켓 상태가 변경되어도 두 쿼리가 모두 호출되는 문제가 수정되었습니다.
