+++ apps/soritor/src/components/provider/QueryProvider.tsx
@@ -5,7 +5,6 @@ const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       throwOnError: true,
-      refetchOnWindowFocus: false,
       retry: false,
       staleTime: 1000 * 60 * 10,
     },
