+++ apps/soritor/src/pages/ticket-list/_components/Deposit.tsx
@@ -1,53 +1,23 @@
 import { Link } from "react-router-dom";
-import { useEffect, useState } from "react";
 import { useToast } from "@uket/ui/components/ui/use-toast";
-import { RefreshCwIcon } from "@uket/ui/components/ui/icon";
 import { Button } from "@uket/ui/components/ui/button";
 
-import Image from "@/components/Image";
-
-import { useQueryTicketQRCode } from "@/hooks/queries/useQueryTicketQRCode";
 import { useQueryDepositUrl } from "@/hooks/queries/useQueryDepositUrl";
 
 import { TicketItem } from "@/types/ticketType";
 
 import { handleCopyClipBoard } from "@/utils/handleCopyToClipboard";
 
-interface QrcodeAndDepositProps {
+interface DepositProps {
   ticketId: TicketItem["ticketId"];
   eventId: TicketItem["eventId"];
   ticketStatus: TicketItem["ticketStatus"];
 }
 
-const QrcodeAndDeposit = (props: QrcodeAndDepositProps) => {
+const Deposit = (props: DepositProps) => {
   const { ticketId, eventId, ticketStatus: isDepositActive } = props;
 
   const { toast } = useToast();
-  const { data: qrcode, refetch } = useQueryTicketQRCode(
-    ticketId,
-    isDepositActive,
-  );
-  const [remainingTime, setRemainingTime] = useState(15); // Initialize countdown state
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setRemainingTime(prev => {
-        if (prev <= 1) {
-          return 0;
-        }
-        return prev - 1; // Decrease countdown
-      });
-    }, 1000);
-
-    return () => clearInterval(timer); // Cleanup on unmount
-  }, []);
-
-  useEffect(() => {
-    if (remainingTime === 0) {
-      refetch(); // 타이머가 0일 때 refetch 호출
-      setRemainingTime(15); // 타이머를 15초로 초기화
-    }
-  }, [remainingTime]);
 
   const { data: deposit } = useQueryDepositUrl(
     ticketId,
@@ -55,42 +25,8 @@ const QrcodeAndDeposit = (props: QrcodeAndDepositProps) => {
     isDepositActive,
   );
 
-  const handleReissueQRCode = () => {
-    refetch();
-    setRemainingTime(15);
-  };
-
   return (
     <>
-      {qrcode && (
-        <>
-          <div>
-            <Image
-              src={qrcode}
-              alt="qrcode"
-              width={100}
-              height={100}
-              className="aspect-square h-36 w-36 scale-125"
-            />
-          </div>
-          <div className="z-50 flex items-center pl-2">
-            <div className="space-x-2">
-              <span>남은시간</span>
-              <span className="text-brand">
-                {remainingTime < 10
-                  ? `00:0${remainingTime}`
-                  : `00:${remainingTime}`}
-              </span>
-            </div>
-            <Button variant="ghost" size="icon" className="rounded-full">
-              <RefreshCwIcon
-                className="h-5 w-5"
-                onClick={handleReissueQRCode}
-              />
-            </Button>
-          </div>
-        </>
-      )}
       {isDepositActive && deposit && (
         <div className="text-center">
           <header className="mb-3 space-y-1.5 font-medium">
@@ -141,4 +77,4 @@ const QrcodeAndDeposit = (props: QrcodeAndDepositProps) => {
   );
 };
 
-export default QrcodeAndDeposit;
+export default Deposit;
